chore: green lint, CI, LF endings, dfxm --version
The 120 tests were not run by anything but me, so CI now runs pytest plus ruff
on Linux and Windows. Making that meaningful meant fixing the lint first:

- Real fixes: h5py groups iterate their keys directly, unused unpack, redundant
  int(round(...)), ClassVar annotations, subprocess check=False, tuple unpacking
  over concatenation, a shadowed loop variable, an unneeded lambda.
- Explicit [tool.ruff.lint] config so CI matches whatever a developer's global
  config happens to be, with the deliberate choices written down: the GUI's
  broad excepts (it must not die on a bad file or a Qt quirk), local imports
  that keep Qt out of the CLI's import path, and Korean UI text that the
  ambiguous-unicode rules flag.

.gitattributes pins LF everywhere, which stops Windows checkouts rewriting
every file and warning about it on each commit.

dfxm --version reads the installed metadata, falling back to a source marker.

+++ src/dfxm/gui/roi.py
@@ -181,6 +181,6 @@ class RectRegionROI(pg.RectROI, AnalysisROI):
     def region_slice(self):
         """(row_slice, col_slice) for indexing the image array, clamped >=0."""
         x, y, wd, ht = self.bounds
-        x0, y0 = int(round(min(x, x + wd))), int(round(min(y, y + ht)))
-        x1, y1 = int(round(max(x, x + wd))), int(round(max(y, y + ht)))
+        x0, y0 = round(min(x, x + wd)), round(min(y, y + ht))
+        x1, y1 = round(max(x, x + wd)), round(max(y, y + ht))
         return slice(max(0, y0), max(0, y1)), slice(max(0, x0), max(0, x1))
